more col top on composite partition fixes

+++ core/src/main/java/io/questdb/cairo/ColumnTopRecorder.java
@@ -59,6 +59,6 @@ public class ColumnTopRecorder implements ColumnTopSink {
 
     @Override
     public void setColumnTop(int columnIndex, long columnTop) {
-        entries.add(columnIndex, columnTop);
+        entries.add((long) columnIndex, columnTop);
     }
 }
